Update V3 - Item Management
Force organisation layout changed, "The Vault" will now be found by
adding it from the "Warband" force selection (helps reduce clutter and
efficient structure.

Fixed bug where magic items and potions were not appearing in characters
options after being selected in the vault.

All items will be set a max limit equal to the number stored in the
vault, this will show an error if the number of a single item has
exceeded what is in the Vault.