Update V4.1
The Big One !
Lots of changes and additions.
The points total is what you total gold will be, so when selecting a
wizard he will add 500 points/ gold to the total.
Selecting anything that would usualy add to the total will reduce it and
0 is no gold left.
Wizard XP points are done by selecting what was achieved as per rules
then adjusting the XP manually.
All magic items, armour, weapons, potions have been added with the
option to buy and sell (You must not go back and change values or reset
them to 0 as the system requires continuity to work efficiently).
Wizard Attributes layout simplified.
Scroll and grimoires are next to be added but are in the works for next
update.